Fix: 회원 탈퇴 시 에러 수정 #121
만든 다이어리가 없는 유저가 회원 탈퇴 할 시 에러 확인 후 수정

+++ src/main/java/com/hanghae/sosohandiary/domain/mypage/service/MyPageService.java
@@ -76,11 +76,15 @@ public class MyPageService {
                 () -> new ApiException(NOT_FOUND_USER)
         );
 
-        Diary diary = diaryRepository.findByMemberId(member.getId()).orElseThrow(
-                () -> new ApiException(NOT_FOUND_DIARY)
-        );
+        List<Diary> diaryList = diaryRepository.findByMemberId(member.getId());
 
-        diaryService.removeDiary(diary.getId(), member);
+        for (Diary diary : diaryList) {
+            if(diaryList.isEmpty()) {
+                diaryService.removeDiary(diary.getId(), member);
+            }
+            diaryService.removeDiary(diary.getId(), member);
+        }
+
         memberRepository.deleteById(member.getId());
         diaryRepository.deleteAllByMemberId(member.getId());
         friendsRepository.deleteAllByFriendId(member.getId());
